requirements.txt
ddddocr==1.4.7
requests_html==0.10.0
requests[socks]==2.25.1
peewee==3.15.3
stem==1.8.0
PyMySQL==1.0.2
retry==0.9.2
pandas==1.3.4
docopt==0.6.2
configparser==5.3.0
termcolor==1.1.0
bs4==0.0.1
openpyxl==3.0.9
Flask==2.2.5
gunicorn==20.1.0
selenium==4.5.0
beautifulsoup4==4.9.3
parse==1.19.0
selenium-wire==5.1.0
pillow==8.4.0
cnocr==2.2.3.2
